big commit, broke out all Products related files into their own folder, refactored all components into one 'Products' component that incorporates the renders,search,filter,and sort. Worked out the interactions between the search and filter features, they now reset eachother. Alphabetized the types array for a better user experience when using the filter feature.
