+++ src/components/App.js
@@ -10,21 +10,12 @@ import {
   Link,
 } from "react-router-dom";
 
-import { getSomething, getAllProducts, getAllTypes } from "../api";
+import { Products } from "./index";
 
-import {
-  Products,
-  ProductSearch,
-  ProductSorter,
-  ProductTypeFilter,
-} from "./index";
+import { getSomething, getAllProducts, getAllTypes } from "../api";
 
 const App = () => {
   const [message, setMessage] = useState("");
-  const [allProducts, setAllProducts] = useState([]);
-  const [currentProducts, setCurrentProducts] = useState([]);
-  const [allTypes, setAllTypes] = useState([]);
-  const [filterMessage, setFilterMessage] = useState("Filter pokemon...");
 
   useEffect(() => {
     getSomething()
@@ -34,45 +25,7 @@ const App = () => {
       .catch((error) => {
         setMessage(error.message);
       });
-
-    getAllProducts()
-      .then((response) => {
-        response.sort((a, b) => {
-          a = a.dex_id;
-          b = b.dex_id;
-          return a - b;
-        });
-        setAllProducts(response);
-        setCurrentProducts(response);
-      })
-      .catch((error) => {
-        console.log("Error fetching products!");
-        console.log(error);
-      });
-
-    getAllTypes()
-      .then((response) => {
-        setAllTypes(response);
-      })
-      .catch((error) => {
-        console.log("Error fetching types!");
-        console.log(error);
-      });
   }, []);
-
-  // type filter function...can this live here??
-  // using it in Products.js and ProductTypeFilter.js
-  function typeFilter(val) {
-    let copy = [...allProducts];
-    let filtered = [];
-    copy.forEach((poke) => {
-      let pokeType = poke.type.toString();
-      if (pokeType.match(val)) {
-        filtered.push(poke);
-      }
-    });
-    setCurrentProducts(filtered);
-  }
 
   return (
     <>
@@ -94,50 +47,7 @@ const App = () => {
             width: "100vw",
           }}
         >
-          <Row
-            style={{
-              marginBottom: "20px",
-              marginTop: "20px",
-              width: "100vw",
-              display: "flex",
-              justifyContent: "center",
-            }}
-          >
-            <ProductSearch
-              allProducts={allProducts}
-              setCurrentProducts={setCurrentProducts}
-              currentProducts={currentProducts}
-            />
-            <ProductTypeFilter
-              allProducts={allProducts}
-              setAllProducts={setAllProducts}
-              currentProducts={currentProducts}
-              setCurrentProducts={setCurrentProducts}
-              allTypes={allTypes}
-              typeFilter={typeFilter}
-              filterMessage={filterMessage}
-              setFilterMessage={setFilterMessage}
-            />
-            <ProductSorter
-              allProducts={allProducts}
-              setAllProducts={setAllProducts}
-              currentProducts={currentProducts}
-              setCurrentProducts={setCurrentProducts}
-            />
-          </Row>
-          <Row
-            style={{
-              width: "100vw",
-              display: "flex",
-              justifyContent: "center",
-            }}
-          >
-            <Products
-              currentProducts={currentProducts}
-              typeFilter={typeFilter}
-              setFilterMessage={setFilterMessage}
-            />
-          </Row>
+          <Products getAllProducts={getAllProducts} getAllTypes={getAllTypes} />
         </Row>
         <Row
           className="bg-secondary"
