+++ src/components/index.js
@@ -1,6 +1,3 @@
 export { default as App } from "./App";
 export { default as ShoppingCart } from "./ShoppingCart";
-export { default as Products } from "./Products";
-export { default as ProductSorter } from "./ProductSorter";
-export { default as ProductTypeFilter } from "./ProductTypeFilter";
-export { default as ProductSearch } from "./ProductSearch";
+export { default as Products } from "./products/Products.js";
