+++ src/components/products/ProductSearch.js
@@ -5,28 +5,17 @@ import { Form } from "react-bootstrap";
 const ProductSearch = ({
   allProducts,
   setCurrentProducts,
-  currentProducts,
+  filterMessage,
+  setFilterMessage,
+  searchVal,
+  setSearchVal,
 }) => {
-  const [searchVal, setSearchVal] = useState("");
-  const [defaultProducts, setDefaultProducts] = useState(allProducts);
-
-  useEffect(() => {
-    console.log("hello");
-    console.log("all", allProducts);
-    console.log("current", currentProducts);
-    console.log("default", defaultProducts);
-    if (searchVal.length === 0) {
-      setDefaultProducts(currentProducts);
-    }
-  }, [searchVal]);
-
+  // search function built to comb the product object for string matches in any of the given fields
   function searcher(val) {
-    let copy = [];
-    if (defaultProducts.length === 0) {
-      copy = [...allProducts];
-    } else {
-      copy = [...defaultProducts];
+    if (filterMessage != "Filter pokemon...") {
+      setFilterMessage("Filter pokemon...");
     }
+    let copy = [...allProducts];
     let filtered = [];
     copy.forEach((poke) => {
       let pokeName = poke.name.toLowerCase();
